organized function order to test the deployment of the cloudformation template
